+++ empire-ui/src/app/(home)/docs/prediction-output-card/page.mdx
@@ -1,0 +1,36 @@
+import { PredictionOutputCardExample } from "@/components/examples/PredictionOutputCardExample";
+import ComponentCodePreview from "@/components/component-code-preview";
+import ComponentInstall from "@/components/component-install";
+import { Breadcrumbs } from "@/components/breadcrumbs";
+
+<Breadcrumbs title={"Card"} />
+
+# Prediction Output Card
+
+---
+
+Displays a prediction output card for an AI model uses framer motion.
+
+<ComponentCodePreview
+  component={<PredictionOutputCardExample />}
+  filePath="examples/PredictionOutputCardExample"
+/>
+
+## Installation
+
+---
+
+<ComponentInstall
+  dependencies={"framer-motion lucide-react @radix-ui/react-tooltip"}
+  componentFilePath={"core/prediction-output-card"}
+  cli={"npx empire@latest add prediction-output-card"}
+/>
+
+## Props
+
+---
+
+| Prop name          | Type     | Description                            |
+| ------------------ | -------- | -------------------------------------- |
+| `length`           | `string` | The length of the passcode.            |
+| `correctPasscode`  | `string` | The correctPasscode.                   |
