+++ empire-ui/src/components/examples/PasscodeCardExample.tsx
@@ -1,6 +1,6 @@
-import { PasscodeCard } from "@/components/core/passcode-input";
+import { PasscodeCard } from "@/components/core/passcode-card";
 
-export const PasscodeCardDemo1 = () => {
+export const PasscodeCardExample = () => {
   return (
     <div className="flex items-center justify-center min-h-screen">
       <PasscodeCard correctPasscode="1234567" length={7} />
